code to turn protoserver on
