+++ src/libZilliqa/CMakeLists.txt
@@ -1,3 +1,3 @@
 add_library (Zilliqa Zilliqa.cpp)
-target_include_directories (Zilliqa PUBLIC ${PROJECT_SOURCE_DIR}/src)
-target_link_libraries (Zilliqa PUBLIC Consensus Crypto DB Lookup Mediator Network Node Server)
+target_include_directories (Zilliqa PUBLIC ${PROJECT_SOURCE_DIR}/src ${CMAKE_BINARY_DIR}/src/libProtoServer ${CMAKE_BINARY_DIR}/src/libMessage)
+target_link_libraries (Zilliqa PUBLIC Consensus Crypto DB Lookup Mediator Network Node Server ProtoServer)
